examples/agent-demo/requirements.txt
# LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# MCP SDK
mcp>=1.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
